requirements.txt
psutil~=7.0.0
click~=8.2.1
python-telegram-bot~=22.1
python-dotenv~=1.1.0
nest-asyncio~=1.6.0